refector: 멀티 모듈 프로젝트로 구조 변경, 기존 CI 파일 삭제
