+++ build.gradle
@@ -4,53 +4,51 @@ plugins {
     id 'io.spring.dependency-management' version '1.1.5'
 }
 
-group = 'com.billing'
-version = '0.0.1-SNAPSHOT'
-
-java {
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(21)
-    }
-}
-
 configurations {
     compileOnly {
         extendsFrom annotationProcessor
     }
 }
 
-repositories {
-    mavenCentral()
+allprojects {
+    group = 'com.billing'
+    version = '0.0.1-SNAPSHOT'
+    sourceCompatibility = '21'
+    repositories {
+        mavenCentral()
+    }
 }
 
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    compileOnly 'org.projectlombok:lombok'
-    runtimeOnly 'com.mysql:mysql-connector-j'
-    annotationProcessor 'org.projectlombok:lombok'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
+subprojects {
+    apply plugin: 'java'
+    apply plugin: 'org.springframework.boot'
+    apply plugin: 'io.spring.dependency-management'
 
-    //JWT
-    implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
-    runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
-    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
-    //JPA
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    //Spring security
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    testImplementation 'org.springframework.security:spring-security-test'
+    dependencies {
+        compileOnly 'org.projectlombok:lombok'
+        runtimeOnly 'com.mysql:mysql-connector-j'
+        annotationProcessor 'org.projectlombok:lombok'
+        testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 
+        implementation 'org.springframework.boot:spring-boot-starter-validation'
+
+        //JPA
+        implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+        //Spring security
+        implementation 'org.springframework.boot:spring-boot-starter-security'
+        testImplementation 'org.springframework.security:spring-security-test'
+
+    }
+
+    tasks.named('test') {
+        useJUnitPlatform()
+    }
+
+    test {
+        jvmArgs '-XX:+EnableDynamicAgentLoading'
+        // 또는
+        // jvmArgs '-Djdk.instrument.traceUsage'
+    }
 }
 
-tasks.named('test') {
-    useJUnitPlatform()
-}
-
-test {
-    jvmArgs '-XX:+EnableDynamicAgentLoading'
-    // 또는
-    // jvmArgs '-Djdk.instrument.traceUsage'
-}
