+++ module-user/build.gradle
@@ -1,0 +1,17 @@
+bootJar {
+    mainClass = 'com.billing.billingProjectApplication'
+}
+
+springBoot {
+    mainClass = 'com.billing.billingProjectApplication'
+}
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+
+    //JWT
+    implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
+    runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
+    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
+}
