+++ module-user/src/main/java/com/billing/BillingProjectApplication.java
@@ -1,0 +1,13 @@
+package com.billing;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BillingProjectApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BillingProjectApplication.class, args);
+	}
+
+}
