+++ module-user/src/main/java/com/billing/user/config/JpaConfig.java
@@ -1,0 +1,10 @@
+package com.billing.user.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+
+}
