+++ module-user/src/main/java/com/billing/user/dto/LoginRequestDTO.java
@@ -1,0 +1,11 @@
+package com.billing.user.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginRequestDTO {
+	private String email;
+	private String password;
+}
