+++ module-user/src/main/java/com/billing/user/dto/ResponseMessage.java
@@ -1,0 +1,14 @@
+package com.billing.user.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+public class ResponseMessage<T> {
+	private Integer statusCode;
+	private String message;
+	private T data;
+}
