+++ module-user/src/main/java/com/billing/user/dto/UserResponseDTO.java
@@ -1,0 +1,27 @@
+package com.billing.user.dto;
+
+import java.time.LocalDateTime;
+
+import com.billing.user.entity.User;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserResponseDTO {
+	private Long id;
+	private String email;
+	private String UserType;
+	private LocalDateTime createdAt;
+	private LocalDateTime modifiedAt;
+
+	public UserResponseDTO(User user) {
+		this.id = user.getId();
+		this.email = user.getEmail();
+		this.UserType = user.getUserType().toString();
+		this.createdAt = user.getCreatedAt();
+		this.modifiedAt = user.getModifiedAt();
+	}
+
+}
