+++ module-user/src/main/java/com/billing/user/entity/UserType.java
@@ -1,0 +1,10 @@
+package com.billing.user.entity;
+
+public enum UserType {
+	USER, //시청자
+	SELEER; //판매자
+
+	public String getAuthority() {
+		return null;
+	}
+}
