+++ module-user/src/main/java/com/billing/user/exception/UserException.java
@@ -1,0 +1,7 @@
+package com.billing.user.exception;
+
+public class UserException extends IllegalArgumentException {
+	public UserException(String message) {
+		super(message);
+	}
+}
