+++ module-user/src/main/java/com/billing/user/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.billing.user.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.billing.user.entity.User;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+	Optional<User> findByEmail(String email);
+}
