+++ settings.gradle
@@ -1,1 +1,5 @@
 rootProject.name = 'billing-project'
+
+include(
+        "module-user"
+)
