--- src/main/java/com/billing/billingproject/BillingProjectApplication.java
@@ -1,13 +1,0 @@
-package com.billing.billingproject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BillingProjectApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BillingProjectApplication.class, args);
-	}
-
-}
