--- src/main/java/com/billing/billingproject/user/config/JpaConfig.java
@@ -1,10 +1,0 @@
-package com.billing.billingproject.user.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-
-}
