--- src/main/java/com/billing/billingproject/user/dto/LoginRequestDTO.java
@@ -1,11 +1,0 @@
-package com.billing.billingproject.user.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginRequestDTO {
-	private String email;
-	private String password;
-}
