--- src/main/java/com/billing/billingproject/user/dto/ResponseMessage.java
@@ -1,14 +1,0 @@
-package com.billing.billingproject.user.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-public class ResponseMessage<T> {
-	private Integer statusCode;
-	private String message;
-	private T data;
-}
