--- src/main/java/com/billing/billingproject/user/entity/UserType.java
@@ -1,10 +1,0 @@
-package com.billing.billingproject.user.entity;
-
-public enum UserType {
-	USER, //시청자
-	SELEER; //판매자
-
-	public String getAuthority() {
-		return null;
-	}
-}
