--- src/main/java/com/billing/billingproject/user/exception/UserException.java
@@ -1,7 +1,0 @@
-package com.billing.billingproject.user.exception;
-
-public class UserException extends IllegalArgumentException {
-	public UserException(String message) {
-		super(message);
-	}
-}
