--- src/main/java/com/billing/billingproject/user/security/UserDetailsServiceImpl.java
@@ -1,25 +1,0 @@
-package com.billing.billingproject.user.security;
-
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import com.billing.billingproject.user.entity.User;
-import com.billing.billingproject.user.repository.UserRepository;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class UserDetailsServiceImpl implements UserDetailsService {
-	private final UserRepository userRepository;
-
-	@Override
-	public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-		User user = userRepository.findByEmail(email)
-			.orElseThrow(() -> new UsernameNotFoundException("Not Found " + email));
-
-		return new UserDetailsImpl(user);
-	}
-}
